feature(#660): correct documentation for breadcrumbs and mark prop activeIndex as int

+++ packages/kotti-ui/source/kotti-breadcrumb/types.ts
@@ -28,7 +28,7 @@ export namespace KottiBreadcrumb {
 	export type Separator = z.infer<typeof separatorSchema>
 
 	export const propsSchema = z.object({
-		activeIndex: z.number().nullable().default(null),
+		activeIndex: z.number().int().nullable().default(null),
 		breadcrumbs: z.array(breadcrumbSchema),
 		separator: separatorSchema.default(
 			() =>
